Fix sme8migration provision failure
The sme8migration provision procedure fails due to
samba-provision.service template expansion. The new unit definition is
not loaded by NSDC systemd instance:

    Warning: samba-provision.service changed on disk. Run
    'systemctl daemon-reload' to reload units.

The fix forces daemon-reload whenever nsdc unit is reloaded, thus the
templates can be expanded at any time.

The fix is not needed if nsdc is restarted or started for the first
time.